Adding roles system

+++ src/main/java/mipt/app/secondmemory/controller/file/FilesControllerImpl.java
@@ -1,5 +1,9 @@
 package mipt.app.secondmemory.controller.file;
 
+import static mipt.app.secondmemory.entity.RoleType.OWNER;
+import static mipt.app.secondmemory.entity.RoleType.READER;
+import static mipt.app.secondmemory.entity.RoleType.WRITER;
+
 import io.minio.Result;
 import io.minio.errors.ErrorResponseException;
 import io.minio.errors.InsufficientDataException;
@@ -9,7 +13,6 @@ import io.minio.errors.ServerException;
 import io.minio.errors.XmlParserException;
 import io.minio.messages.Item;
 import jakarta.servlet.http.Part;
-
 import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
@@ -20,14 +23,22 @@ import mipt.app.secondmemory.dto.directory.DirectoryInfoRequest;
 import mipt.app.secondmemory.dto.directory.RootDirectoriesRequest;
 import mipt.app.secondmemory.dto.file.FileInfoRequest;
 import mipt.app.secondmemory.dto.file.FileInfoResponse;
+import mipt.app.secondmemory.entity.Role;
+import mipt.app.secondmemory.entity.Session;
+import mipt.app.secondmemory.exception.directory.BucketNotFoundException;
 import mipt.app.secondmemory.exception.directory.NoSuchBucketException;
 import mipt.app.secondmemory.exception.directory.NoSuchDirectoryException;
 import mipt.app.secondmemory.exception.file.DatabaseException;
 import mipt.app.secondmemory.exception.file.FileAlreadyExistsException;
 import mipt.app.secondmemory.exception.file.FileMemoryLimitExceededException;
 import mipt.app.secondmemory.exception.file.FileNotFoundException;
+import mipt.app.secondmemory.exception.role.NoRoleFoundException;
+import mipt.app.secondmemory.exception.session.SessionNotFoundException;
+import mipt.app.secondmemory.repository.RolesRepository;
+import mipt.app.secondmemory.repository.SessionsRepository;
 import mipt.app.secondmemory.service.FilesService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -37,9 +48,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 public class FilesControllerImpl implements FilesController {
   private final FilesService filesService;
+  private final RolesRepository rolesRepository;
+  private final SessionsRepository sessionsRepository;
 
   @Override
-  public ResponseEntity<FileInfoResponse> uploadFiles(String bucketName, Part file)
+  public ResponseEntity<FileInfoResponse> uploadFile(Long bucketId, Part file, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           ErrorResponseException,
@@ -50,13 +63,17 @@ public class FilesControllerImpl implements FilesController {
           NoSuchBucketException,
           InvalidResponseException,
           XmlParserException,
-          InternalException {
-
-    return ResponseEntity.ok(filesService.uploadFile(bucketName, file));
+          InternalException,
+          BucketNotFoundException,
+          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    return ResponseEntity.ok(filesService.uploadFile(bucketId, file, session.getUser()));
   }
 
   @Override
-  public ResponseEntity<Void> renameFile(String bucketName, String oldKey, String newKey)
+  public ResponseEntity<Void> renameFile(
+      String bucketName, Long fileId, String oldKey, String newKey, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -66,13 +83,24 @@ public class FilesControllerImpl implements FilesController {
           InvalidKeyException,
           InvalidResponseException,
           XmlParserException,
-          InternalException {
+          InternalException,
+          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != WRITER && userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to rename file");
+    }
     filesService.renameFile(bucketName, oldKey, newKey);
     return ResponseEntity.ok().build();
   }
 
   @Override
-  public ResponseEntity<Void> deleteFile(String bucketName, String key)
+  public ResponseEntity<Void> deleteFile(
+      String bucketName, Long fileId, String key, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -82,14 +110,30 @@ public class FilesControllerImpl implements FilesController {
           InvalidKeyException,
           InvalidResponseException,
           XmlParserException,
-          InternalException {
+          InternalException,
+          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to delete file");
+    }
     filesService.deleteFile(bucketName, key);
     return ResponseEntity.ok().build();
   }
 
   @Override
   public ResponseEntity<Void> moveFile(
-      String oldBucketName, String newBucketName, String fileName, String oldPath, String newPath)
+      String oldBucketName,
+      String newBucketName,
+      String fileName,
+      String oldPath,
+      String newPath,
+      Long fileId,
+      String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -101,9 +145,35 @@ public class FilesControllerImpl implements FilesController {
           XmlParserException,
           InternalException,
           FileAlreadyExistsException,
-          NoSuchBucketException {
+          NoSuchBucketException,
+          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to move file");
+    }
     filesService.moveFile(oldBucketName, newBucketName, fileName, oldPath, newPath);
     return ResponseEntity.ok().build();
+  }
+
+  @Override
+  public ResponseEntity<FileInfoResponse> uploadFileToFolder(Long folderId, Part file)
+      throws NoSuchDirectoryException,
+          BucketNotFoundException,
+          ServerException,
+          InsufficientDataException,
+          ErrorResponseException,
+          IOException,
+          NoSuchAlgorithmException,
+          InvalidKeyException,
+          InvalidResponseException,
+          XmlParserException,
+          InternalException {
+    return ResponseEntity.ok(filesService.uploadFileToFolder(folderId, file));
   }
 
   @Override
@@ -112,20 +182,34 @@ public class FilesControllerImpl implements FilesController {
   }
 
   @Override
-  public ResponseEntity<FileInfoResponse> getFileInfo(long fileId, FileInfoRequest fileInfoRequest)
-      throws FileNotFoundException, DatabaseException {
+  public ResponseEntity<FileInfoResponse> getFileInfo(
+      long fileId, FileInfoRequest fileInfoRequest, String cookieValue)
+      throws FileNotFoundException, DatabaseException, SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != READER
+        && userRole.getType() != WRITER
+        && userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to read file information");
+    }
     return ResponseEntity.ok(filesService.getFileInfo(fileId, fileInfoRequest));
   }
 
   @Override
   public ResponseEntity<Iterable<Result<Item>>> getFilesInDirectory(
-      DirectoryInfoRequest directoryInfoRequest) throws NoSuchDirectoryException {
+      DirectoryInfoRequest directoryInfoRequest, String cookieValue)
+      throws NoSuchDirectoryException {
     return ResponseEntity.ok(filesService.getFilesInDirectory(directoryInfoRequest));
   }
 
   @Override
   public ResponseEntity<Iterable<Result<Item>>> getRootDirectories(
-      RootDirectoriesRequest rootDirectoriesRequest) throws NoSuchBucketException {
+      RootDirectoriesRequest rootDirectoriesRequest, String cookieValue)
+      throws NoSuchBucketException {
     return ResponseEntity.ok(filesService.getRootDirectories(rootDirectoriesRequest));
   }
 }
